Revert "#78 broken implementation with DockerImageReferenceFactory"
This reverts commit 29d3872a
