+++ src/test/java/com/exasol/containers/ExasolDockerImageReferenceTest.java
@@ -1,4 +1,4 @@
-package com.exasol.containers.imagereference;
+package com.exasol.containers;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -6,7 +6,7 @@ import static org.hamcrest.Matchers.equalTo;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class DockerImageReferenceFactoryTest {
+class ExasolDockerImageReferenceTest {
     @CsvSource({ //
             "7, exasol/docker-db:7.0.0-d1", //
             "6.1, exasol/docker-db:6.1.0-d1", //
@@ -14,10 +14,9 @@ class DockerImageReferenceFactoryTest {
             "7.0.4-d2, exasol/docker-db:7.0.4-d2", //
             "111.222.333-d444, exasol/docker-db:111.222.333-d444", //
             "foo/bar:latest, foo/bar:latest", //
-            "baz/zoo:1.2.3.4, baz/zoo:1.2.3.4", //
-            "exasol/docker-db:7, exasol/docker-db:7.0.0-d1" })
+            "baz/zoo:1.2.3.4, baz/zoo:1.2.3.4" })
     @ParameterizedTest
     void toString(final String input, final String expectedReference) {
-        assertThat(DockerImageReferenceFactory.getInstance().parse(input).toString(), equalTo(expectedReference));
+        assertThat(ExasolDockerImageReference.parse(input).toString(), equalTo(expectedReference));
     }
 }
